requirements.txt
fastapi
uvicorn[standard]
motor
beanie
python-jose
passlib[bcrypt]
python-dotenv
email-validator
bcrypt<4.0
python-jose[cryptography]
